updated fontstash and reorganized folder structure so that it compiles with OF 0.8.0 Makefile system
