+++ libs/fontstash/include/fontstash.h
@@ -17,7 +17,6 @@
 // 3. This notice may not be removed or altered from any source distribution.
 //
 
-
 #ifndef FONTSTASH_H
 #define FONTSTASH_H
 
@@ -26,10 +25,10 @@ struct sth_stash* sth_create(int cachew, int cacheh);
 int sth_add_font(struct sth_stash* stash, const char* path);
 int sth_add_font_from_memory(struct sth_stash* stash, unsigned char* buffer);
 
-int sth_add_bitmap_font(struct sth_stash* stash, int ascent, int descent, int line_gap);
-int sth_add_glyph(struct sth_stash* stash, int idx, GLuint id, const char* s,
-                  short size, short base, int x, int y, int w, int h,
-                  float xoffset, float yoffset, float xadvance);
+int  sth_add_bitmap_font(struct sth_stash* stash, int ascent, int descent, int line_gap);
+void sth_add_glyph(struct sth_stash* stash, int idx, GLuint id, const char* s,  /* @rlyeh: function does not return int */
+                   short size, short base, int x, int y, int w, int h,
+                   float xoffset, float yoffset, float xadvance);
 
 void sth_begin_draw(struct sth_stash* stash);
 void sth_end_draw(struct sth_stash* stash);
